requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.10
django-allauth==0.44.0
django-cors-headers==3.10.1
django-environ==0.8.1
django-heroku==0.0.0
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
greenlet==1.1.2
gunicorn==19.9.0
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.0
oauthlib==3.1.1
pandas==1.3.5
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
zope.event==4.5.0
zope.interface==5.4.0
